teams/team: remove container
